using unsafe to initialize array, as Standingorder does not implement Copy trait

+++ driver/src/models/mod.rs
@@ -1,6 +1,7 @@
 pub mod state;
 pub mod flux;
 pub mod order;
+#[macro_use]
 pub mod standing_order;
 
 pub use crate::models::state::State;
@@ -13,7 +14,7 @@ use web3::types::H256;
 
 //ToDo: get variables from database
 pub const TOKENS: u8 = 30;
-pub const NUM_RESERVED_ACCOUNTS: u8 = 50;
+pub const NUM_RESERVED_ACCOUNTS: usize = 50;
 pub const DB_NAME: &str = "dfusion2";
 
 pub trait RollingHashable {
